+++ src/main/java/com/honeyboard/api/project/finale/model/request/FinaleProjectCreate.java
@@ -1,4 +1,4 @@
-package com.honeyboard.api.project.finale.model.Request;
+package com.honeyboard.api.project.finale.model.request;
 
 import lombok.*;
 
@@ -8,10 +8,11 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public
-class FinaleProjectUpdate {
+class FinaleProjectCreate {
 
     String title;
     String description;
     String url;
+    TeamRequest teams;
 
 }
