+++ src/main/java/com/honeyboard/api/project/finale/model/response/FinaleProjectList.java
@@ -1,4 +1,4 @@
-package com.honeyboard.api.project.finale.model.Response;
+package com.honeyboard.api.project.finale.model.response;
 
 import com.honeyboard.api.project.model.ProjectUserInfo;
 import lombok.*;
